Cleaned up warnings and errors

+++ processes/getty/src/getty.c
@@ -46,7 +46,7 @@ int main()
             tcsetpgrp(0, 0);
             execve("shell", 0, 0);
         }
-        uint64_t status;
+        int status;
         waitpid(pid, &status, 0);
         tcsetpgrp(0, 0);
     }
